LASB-3091 - refactored code.
